+++ verik-compiler/src/main/kotlin/io/verik/compiler/main/VerikCompilerException.kt
@@ -8,4 +8,4 @@ package io.verik.compiler.main
  * Exception to be thrown when the compiler terminates with an error status. This should be thrown by the message
  * collector and should not be thrown directly.
  */
-class VerikException : Exception()
+class VerikCompilerException : Exception()